+++ be/src/pipeline/exec/hashjoin_probe_operator.cpp
@@ -17,6 +17,7 @@
 
 #include "hashjoin_probe_operator.h"
 
+#include <cstdint>
 #include <string>
 
 #include "common/logging.h"
@@ -644,7 +645,7 @@ Status HashJoinProbeOperatorX::prepare(RuntimeState* state) {
         }
     }
 
-    const int right_col_idx =
+    const int64_t right_col_idx =
             (_is_right_semi_anti && !_have_other_join_conjunct) ? 0 : _left_table_data_types.size();
     size_t idx = 0;
     for (const auto* slot : slots_to_check) {
