+++ be/src/pipeline/exec/nested_loop_join_build_operator.cpp
@@ -109,9 +109,9 @@ Status NestedLoopJoinBuildSinkOperatorX::init(const TPlanNode& tnode, RuntimeSta
 
 Status NestedLoopJoinBuildSinkOperatorX::prepare(RuntimeState* state) {
     // pre-compute the tuple index of build tuples in the output row
-    int num_build_tuples = _child_x->row_desc().tuple_descriptors().size();
+    int64_t num_build_tuples = _child_x->row_desc().tuple_descriptors().size();
 
-    for (int i = 0; i < num_build_tuples; ++i) {
+    for (int64_t i = 0; i < num_build_tuples; ++i) {
         TupleDescriptor* build_tuple_desc = _child_x->row_desc().tuple_descriptors()[i];
         auto tuple_idx = _row_descriptor.get_tuple_idx(build_tuple_desc->id());
         RETURN_IF_INVALID_TUPLE_IDX(build_tuple_desc->id(), tuple_idx);
