+++ be/src/vec/common/sort/partition_sorter.h
@@ -74,11 +74,12 @@ class PartitionSorter final : public Sorter {
     ENABLE_FACTORY_CREATOR(PartitionSorter);
 
 public:
-    PartitionSorter(VSortExecExprs& vsort_exec_exprs, int limit, int64_t offset, ObjectPool* pool,
-                    std::vector<bool>& is_asc_order, std::vector<bool>& nulls_first,
-                    const RowDescriptor& row_desc, RuntimeState* state, RuntimeProfile* profile,
-                    bool has_global_limit, int partition_inner_limit,
-                    TopNAlgorithm::type top_n_algorithm, SortCursorCmp* previous_row);
+    PartitionSorter(VSortExecExprs& vsort_exec_exprs, int64_t limit, int64_t offset,
+                    ObjectPool* pool, std::vector<bool>& is_asc_order,
+                    std::vector<bool>& nulls_first, const RowDescriptor& row_desc,
+                    RuntimeState* state, RuntimeProfile* profile, bool has_global_limit,
+                    int64_t partition_inner_limit, TopNAlgorithm::type top_n_algorithm,
+                    SortCursorCmp* previous_row);
 
     ~PartitionSorter() override = default;
 
@@ -100,7 +101,7 @@ private:
     int64 _output_total_rows = 0;
     int64 _output_distinct_rows = 0;
     bool _has_global_limit = false;
-    int _partition_inner_limit = 0;
+    int64_t _partition_inner_limit = 0;
     TopNAlgorithm::type _top_n_algorithm = TopNAlgorithm::type::ROW_NUMBER;
     SortCursorCmp* _previous_row = nullptr;
 };
