+++ be/src/vec/exprs/vectorized_agg_fn.h
@@ -97,7 +97,7 @@ public:
     bool is_merge() const { return _is_merge; }
     const VExprContextSPtrs& input_exprs_ctxs() const { return _input_exprs_ctxs; }
 
-    static Status check_agg_fn_output(int key_size,
+    static Status check_agg_fn_output(int64_t key_size,
                                       const std::vector<vectorized::AggFnEvaluator*>& agg_fn,
                                       const RowDescriptor& output_row_desc);
